requirements.txt
langchain-community
langchain
qdrant-client
pymupdf
rapidocr-onnxruntime
pypdf
fastembed
torch
transformers
langchain-huggingface
python-dotenv
streamlit
streamlit-chat